+++ R/plotting.R
@@ -999,6 +999,7 @@ SplitDotPlotGG <- function(
 #' @param nCol Number of columns to use when plotting multiple features.
 #' @param no.axes Remove axis labels
 #' @param no.legend Remove legend from the graph. Default is TRUE.
+#' @param coord.fixed Use a fixed scale coordinate system (for spatial coordinates). Default is FALSE.
 #' @param dark.theme Plot in a dark theme
 #' @param do.return return the ggplot2 object
 #' @param vector.friendly FALSE by default. If TRUE, points are flattened into a PNG, while axes/labels retain full vector resolution. Useful for producing AI-friendly plots with large numbers of cells.
@@ -1032,6 +1033,7 @@ FeaturePlot <- function(
   nCol = NULL,
   no.axes = FALSE,
   no.legend = TRUE,
+  coord.fixed = FALSE,
   dark.theme = FALSE,
   do.return = FALSE,
   vector.friendly=FALSE
@@ -1123,6 +1125,7 @@ FeaturePlot <- function(
         dim.codes = dim.codes,
         min.cutoff = min.cutoff,
         max.cutoff = max.cutoff,
+        coord.fixed = coord.fixed,
         no.axes = no.axes,
         no.legend = no.legend,
         dark.theme = dark.theme
@@ -1135,6 +1138,7 @@ FeaturePlot <- function(
       feature = features.plot,
       min.cutoff = min.cutoff,
       max.cutoff = max.cutoff,
+      coord.fixed = coord.fixed,
       MoreArgs = list( # Arguments that are not being repeated
         data.use = data.use,
         data.plot = data.plot,
@@ -1293,6 +1297,7 @@ FeatureHeatmap <- function(
   ))
   colnames(x = data.plot)[1:2] <- c("dim1", "dim2")
   data.plot$ident <- as.character(x = object@ident)
+  data.plot <- data.plot[data.plot$ident %in% idents.use,] # keep only identities defined in idents.use
   data.plot$cell <- rownames(x = data.plot)
   features.plot <- gsub('-', '\\.', features.plot)
   data.plot  %>% gather(key = "gene", value = "expression", -dim1, -dim2, -ident, -cell) -> data.plot
@@ -1544,7 +1549,9 @@ globalVariables(names = 'Value', package = 'Seurat', add = TRUE)
 #' @param plot.x.lim X-axis maximum on each QQ plot.
 #' @param plot.y.lim Y-axis maximum on each QQ plot.
 #'
-#' @return A ggplot object
+#' @return Returns a Seurat object where object@@dr$pca@@jackstraw@@overall.p.values
+#' represents p-values for each PC and object@@dr$pca@@misc$jackstraw.plot 
+#' stores the ggplot2 plot.
 #'
 #' @author Thanks to Omri Wurtzel for integrating with ggplot
 #'
@@ -1569,7 +1576,7 @@ JackStrawPlot <- function(
   pAll <- pAll[, PCs, drop = FALSE]
   pAll <- as.data.frame(pAll)
   pAll$Contig <- rownames(x = pAll)
-  pAll.l <- melt(data = pAll, id.vars = "Contig")
+  pAll.l <- reshape2::melt(data = pAll, id.vars = "Contig")
   colnames(x = pAll.l) <- c("Contig", "PC", "Value")
   qq.df <- NULL
   score.df <- NULL
@@ -1620,7 +1627,11 @@ JackStrawPlot <- function(
     coord_flip() +
     geom_abline(intercept = 0, slope = 1, linetype = "dashed", na.rm = TRUE) +
     theme_bw()
-  return(gp)
+  
+  object@dr$pca@misc[["jackstraw.plot"]] <- gp
+  print(gp)
+  
+  return(object)
 }
 
 globalVariables(names = c('x', 'y'), package = 'Seurat', add = TRUE)
@@ -2353,6 +2364,7 @@ globalVariables(names = c('x', 'y', 'ident'), package = 'Seurat', add = TRUE)
 #' @param do.label Whether to label the clusters
 #' @param label.size Sets size of labels
 #' @param no.legend Setting to TRUE will remove the legend
+#' @param coord.fixed Use a fixed scale coordinate system (for spatial coordinates). Default is FALSE.
 #' @param no.axes Setting to TRUE will remove the axes
 #' @param dark.theme Use a dark theme for the plot
 #' @param plot.order Specify the order of plotting for the idents. This can be
@@ -2412,6 +2424,7 @@ DimPlot <- function(
   do.label = FALSE,
   label.size = 4,
   no.legend = FALSE,
+  coord.fixed = FALSE,
   no.axes = FALSE,
   dark.theme = FALSE,
   plot.order = NULL,
@@ -2486,7 +2499,9 @@ DimPlot <- function(
   data.plot$pt.size <- pt.size
   if (!is.null(x = cells.highlight)) {
     # Ensure that cells.highlight are in our data.frame
-    if (is.data.frame(x = cells.highlight) || !is.list(x = cells.highlight)) {
+    if (is.character(x = cells.highlight)) {
+      cells.highlight <- list(cells.highlight)
+    } else if (is.data.frame(x = cells.highlight) || !is.list(x = cells.highlight)) {
       cells.highlight <- as.list(x = cells.highlight)
     }
     cells.highlight <- lapply(
@@ -2522,6 +2537,9 @@ DimPlot <- function(
         length.out = length(x = cells.highlight)
       )
       highlight <- rep_len(x = NA_character_, length.out = nrow(x = data.plot))
+      if (is.null(x = cols.use)) {
+        cols.use <- 'black'
+      }
       cols.use <- c(cols.use[1], cols.highlight)
       size <- rep_len(x = pt.size, length.out = nrow(x = data.plot))
       for (i in 1:length(x = cells.highlight)) {
@@ -2569,6 +2587,9 @@ DimPlot <- function(
   }
   if (!is.null(x = cols.use)) {
     p <- p + scale_colour_manual(values = cols.use, na.value=na.value)
+  }
+  if(coord.fixed){
+    p <- p + coord_fixed()
   }
   p <- p + guides(size = FALSE)
   p2 <- p +