+++ tests/testthat/test_alignment.R
@@ -31,3 +31,9 @@ test_that("CalcVarExpRatio performs as expectd", {
   expect_equal(pbmc_cca@meta.data$var.ratio.pca[40], 0.5198426, tolerance = 1e-6)
   expect_equal(pbmc_cca@meta.data$var.ratio.pca[80], 0.9824946, tolerance = 1e-6)
 })
+
+test_that("RunMultiCCA works with add.cell.ids", {
+  pbmc_multi_cca <- RunMultiCCA(list(pbmc_small, pbmc_small, pbmc_small),
+                                add.cell.ids = c("A", "B", "C"))
+  expect_s4_class(pbmc_multi_cca, "seurat")
+})