+++ NEWS.md
@@ -2,6 +2,12 @@
 All notable changes to Seurat will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## Develop
+### Added
+- RenameCells added for easy renaming of all cells
+- Read10X_h5 added to read in 10X formatted h5 files
+- SetAssayData ensures cell order is the same between assay objects and the Seurat object
+- Compatability updates for ggplot2 v2.3.0
 
 ## [2.3.1] - 2018-05-03
 ### Added